Reorganise and document internal types
