+++ src/reader.rs
@@ -8,11 +8,12 @@ use std::num::NonZeroUsize;
 use heed::RoTxn;
 use ordered_float::OrderedFloat;
 
+use crate::distance::Distance;
+use crate::internals::{KeyCodec, Side};
 use crate::item_iter::ItemIter;
 use crate::node::{Descendants, ItemIds, Leaf, SplitPlaneNormal, UnalignedF32Slice};
 use crate::{
-    Database, Distance, Error, ItemId, Key, KeyCodec, MetadataCodec, Node, NodeId, Prefix,
-    PrefixCodec, Result, Side,
+    Database, Error, ItemId, Key, MetadataCodec, Node, NodeId, Prefix, PrefixCodec, Result,
 };
 
 #[derive(Debug)]
